Modularized app icon code into functions in a helper script
